cleanup: vector rotate verification tests
